+++ packages/api/src/apps/api/app.ts
@@ -35,8 +35,6 @@ export default async (env: ProcessEnv) => {
   const detectContractsUpdateRateS = Number(env.DETECT_CONTRACTS_UPDATE_RATE_S || 60);
   // Defaults to 15 minutes, prices dont update in coingecko or other calls very fast
   const priceUpdateRateS = Number(env.PRICE_UPDATE_RATE_S || 15 * 60);
-  // default to 10 days worth of blocks
-  const oldestBlock = Number(env.OLDEST_BLOCK_MS || 10 * 60 * 60 * 24 * 1000);
 
   assert(updateRateS >= 1, "UPDATE_RATE_S must be 1 or higher");
   assert(detectContractsUpdateRateS >= 1, "DETECT_CONTRACTS_UPDATE_RATE_S must be 1 or higher");
@@ -51,7 +49,6 @@ export default async (env: ProcessEnv) => {
     web3,
     coingecko: new Coingecko(),
     zrx: new Zrx(env.zrxBaseUrl),
-    blocks: tables.blocks.Table(),
     emps: {
       active: tables.emps.Table("Active Emp"),
       expired: tables.emps.Table("Expired Emp"),
@@ -128,7 +125,6 @@ export default async (env: ProcessEnv) => {
   // services for ingesting data
   const services = {
     // these services can optionally be configured with a config object, but currently they are undefined or have defaults
-    blocks: Services.Blocks(undefined, appState),
     emps: Services.EmpState({ debug }, appState),
     registry: await Services.Registry({ debug }, appState, (event, data) =>
       serviceEvents.emit("empRegistry", event, data)
@@ -233,12 +229,10 @@ export default async (env: ProcessEnv) => {
     // ignore case when startblock == endblock, this can happen when loop is run before a new block has changed
     if (startBlock === endBlock) return;
     assert(startBlock < endBlock, "Startblock must be lower than endBlock");
-    await services.blocks.handleNewBlock(endBlock);
     // update everyting
     await services.emps.update(startBlock, endBlock);
     await services.lsps.update(startBlock, endBlock);
     await services.erc20s.update();
-    await services.blocks.cleanBlocks(oldestBlock);
     appState.lastBlockUpdate = endBlock;
   }