+++ packages/api/src/apps/datastore_api/app.ts
@@ -1,27 +1,24 @@
 import assert from "assert";
 import { ethers } from "ethers";
+import moment from "moment";
 import Events from "events";
-import Web3 from "web3";
-
 import { tables, Coingecko, utils, Multicall2 } from "@uma/sdk";
+import { Datastore } from "@google-cloud/datastore";
 
 import * as Services from "../../services";
 import Express from "../../services/express-channels";
 import * as Actions from "../../services/actions";
 import { ProcessEnv, AppState, Channels } from "../../types";
-import { empStats, empStatsHistory, lsps } from "../../tables";
+import { empStats, empStatsHistory, lsps, StoresFactory } from "../../tables";
 import Zrx from "../../libs/zrx";
-import { Profile, parseEnvArray, BlockInterval, expirePromise } from "../../libs/utils";
+import { Profile, parseEnvArray, getWeb3, BlockInterval, expirePromise } from "../../libs/utils";
 
-// This is almost identical to api app, but removes some key services from starting up. Maintains ability to use
-// api for LSP calls only. Express API maintains compatibility with original API but will not populate EMP data.
-// Eventually much of the code in this file could be refactored to deduplicate a lot of the shared code with API
-// but would require more tooling/thought around how to turn this into configuration rather than code.
 export default async (env: ProcessEnv) => {
   assert(env.CUSTOM_NODE_URL, "requires CUSTOM_NODE_URL");
   assert(env.EXPRESS_PORT, "requires EXPRESS_PORT");
   assert(env.zrxBaseUrl, "requires zrxBaseUrl");
   assert(env.MULTI_CALL_2_ADDRESS, "requires MULTI_CALL_2_ADDRESS");
+  assert(env.GOOGLE_APPLICATION_CREDENTIALS, "requires GOOGLE_APPLICATION_CREDENTIALS");
   const lspCreatorAddresses = parseEnvArray(env.lspCreatorAddresses || "");
 
   // debug flag for more verbose logs
@@ -30,8 +27,11 @@ export default async (env: ProcessEnv) => {
 
   const provider = new ethers.providers.WebSocketProvider(env.CUSTOM_NODE_URL);
 
-  // how often to run expensive state updates, defaults to 1 minutes since EMP updates are gone
-  const updateRateS = Number(env.UPDATE_RATE_S || 60);
+  // we need web3 for syth price feeds
+  const web3 = getWeb3(env.CUSTOM_NODE_URL);
+
+  // how often to run expensive state updates, defaults to 10 minutes
+  const updateRateS = Number(env.UPDATE_RATE_S || 10 * 60);
   // Defaults to 60 seconds, since this is i think a pretty cheap call, and we want to see new contracts quickly
   const detectContractsUpdateRateS = Number(env.DETECT_CONTRACTS_UPDATE_RATE_S || 60);
   // Defaults to 15 minutes, prices dont update in coingecko or other calls very fast
@@ -44,16 +44,17 @@ export default async (env: ProcessEnv) => {
   // services can emit events when necessary, though for now any services that depend on events must be in same process
   const serviceEvents = new Events();
 
+  const datastoreClient = new Datastore();
+  const datastores = StoresFactory(datastoreClient);
   // state shared between services
   const appState: AppState = {
     provider,
-    // This isnt needed for the lsp app to run, but the type needs to conform to app state
-    web3: {} as Web3,
+    web3,
     coingecko: new Coingecko(),
     zrx: new Zrx(env.zrxBaseUrl),
     emps: {
-      active: tables.emps.Table("Active Emp"),
-      expired: tables.emps.Table("Expired Emp"),
+      active: tables.emps.Table("Active Emp", datastores.empsActive),
+      expired: tables.emps.Table("Expired Emp", datastores.empsExpired),
     },
     prices: {
       usd: {
@@ -68,31 +69,31 @@ export default async (env: ProcessEnv) => {
     marketPrices: {
       usdc: {
         latest: {},
-        history: empStatsHistory.Table("Market Price"),
+        history: empStatsHistory.Table("Market Price", datastores.empStatsHistory),
       },
     },
-    erc20s: tables.erc20s.Table(),
+    erc20s: tables.erc20s.Table("Erc20", datastores.erc20),
     stats: {
       emp: {
         usd: {
           latest: {
-            tvm: empStats.Table("Latest Tvm"),
-            tvl: empStats.Table("Latest Tvl"),
+            tvm: empStats.Table("Latest Tvm", datastores.empStatsTvm),
+            tvl: empStats.Table("Latest Tvl", datastores.empStatsTvl),
           },
           history: {
-            tvm: empStatsHistory.Table("Tvm History"),
-            tvl: empStatsHistory.Table("Tvl History"),
+            tvm: empStatsHistory.Table("Tvm History", datastores.empStatsTvlHistory),
+            tvl: empStatsHistory.Table("Tvl History", datastores.empStatsTvmHistory),
           },
         },
       },
       lsp: {
         usd: {
           latest: {
-            tvl: empStats.Table("Latest Tvl"),
-            tvm: empStats.Table("Latest Tvm"),
+            tvl: empStats.Table("Latest Tvl", datastores.lspStatsTvl),
+            tvm: empStats.Table("Latest Tvm", datastores.lspStatsTvm),
           },
           history: {
-            tvl: empStatsHistory.Table("Tvl History"),
+            tvl: empStatsHistory.Table("Tvl History", datastores.lspStatsTvlHistory),
           },
         },
       },
@@ -119,8 +120,8 @@ export default async (env: ProcessEnv) => {
     shortAddresses: new Set<string>(),
     multicall2: new Multicall2(env.MULTI_CALL_2_ADDRESS, provider),
     lsps: {
-      active: lsps.Table("Active LSP"),
-      expired: lsps.Table("Expired LSP"),
+      active: lsps.Table("Active LSP", datastores.lspsActive),
+      expired: lsps.Table("Expired LSP", datastores.lspsExpired),
     },
   };
 
@@ -132,6 +133,16 @@ export default async (env: ProcessEnv) => {
       serviceEvents.emit("empRegistry", event, data)
     ),
     collateralPrices: Services.CollateralPrices({ debug }, appState),
+    syntheticPrices: Services.SyntheticPrices(
+      {
+        debug,
+        cryptowatchApiKey: env.cryptowatchApiKey,
+        tradermadeApiKey: env.tradermadeApiKey,
+        quandlApiKey: env.quandlApiKey,
+        defipulseApiKey: env.defipulseApiKey,
+      },
+      appState
+    ),
     erc20s: Services.Erc20s({ debug }, appState),
     empStats: Services.stats.Emp({ debug }, appState),
     marketPrices: Services.MarketPrices({ debug }, appState),
@@ -145,8 +156,15 @@ export default async (env: ProcessEnv) => {
 
   const initBlock = await provider.getBlock("latest");
 
+  // warm caches
+  await services.registry(appState.lastBlockUpdate, initBlock.number);
+  console.log("Got all EMP addresses");
+
   await services.lspCreator.update(appState.lastBlockUpdate, initBlock.number);
   console.log("Got all LSP addresses");
+
+  await services.emps.update(appState.lastBlockUpdate, initBlock.number);
+  console.log("Updated EMP state");
 
   await services.lsps.update(appState.lastBlockUpdate, initBlock.number);
   console.log("Updated LSP state");
@@ -157,14 +175,35 @@ export default async (env: ProcessEnv) => {
   await services.erc20s.update();
   console.log("Updated tokens");
 
+  // backfill price histories, disable if not specified in env
+  if (env.backfillDays) {
+    console.log(`Backfilling price history from ${env.backfillDays} days ago`);
+    await services.collateralPrices.backfill(moment().subtract(env.backfillDays, "days").valueOf());
+    console.log("Updated Collateral Prices Backfill");
+    await services.empStats.backfill();
+    console.log("Updated EMP Backfill");
+
+    await services.lspStats.backfill();
+    console.log("Updated LSP Backfill");
+  }
+
   await services.collateralPrices.update();
   console.log("Updated Collateral Prices");
+
+  await services.syntheticPrices.update();
+  console.log("Updated Synthetic Prices");
+
+  await services.empStats.update();
+  console.log("Updated EMP Stats");
 
   await services.lspStats.update();
   console.log("Updated LSP Stats");
 
   await services.globalStats.update();
   console.log("Updated Global Stats");
+
+  await services.marketPrices.update();
+  console.log("Updated Market Prices");
 
   // services consuming data
   const channels: Channels = [
@@ -193,6 +232,8 @@ export default async (env: ProcessEnv) => {
     // ignore case when startblock == endblock, this can happen when loop is run before a new block has changed
     if (startBlock === endBlock) return;
     assert(startBlock < endBlock, "Startblock must be lower than endBlock");
+    // update everyting
+    await services.emps.update(startBlock, endBlock);
     await services.lsps.update(startBlock, endBlock);
     await services.erc20s.update();
     appState.lastBlockUpdate = endBlock;
@@ -201,83 +242,83 @@ export default async (env: ProcessEnv) => {
   // separate out price updates into a different loop to query every few minutes
   async function updatePrices() {
     await services.collateralPrices.update();
+    await services.syntheticPrices.update();
     await services.marketPrices.update();
+    await services.empStats.update();
     await services.lspStats.update();
     await services.globalStats.update();
   }
 
   // wait update rate before running loops, since all state was just updated on init
-  new Promise((res) => setTimeout(res, updateRateS * 1000)).then(initLoops);
+  await new Promise((res) => setTimeout(res, updateRateS * 1000));
 
-  async function initLoops() {
-    console.log("Starting LSP API update loops");
+  console.log("Starting API update loops");
 
-    // listen for new lsp contracts since after we have started api, and make sure they get state updated asap
-    // These events should only be bound after startup, since initialization above takes care of updating all contracts on startup
-    // Because there is now a event driven dependency, the lsp creator and lsp state updater must be in same process
-    serviceEvents.on("multiLspCreator", (event, data) => {
-      // handle created events
-      if (event === "created") {
-        console.log("LspCreator found a new contract", JSON.stringify(data));
-        services.lsps.updateLsps([data.address], data.startBlock, data.endBlock).catch(console.error);
-      }
-    });
-
-    // listen for new emp contracts since after we start api, make sure they get state updates asap
-    serviceEvents.on("empRegistry", (event, data) => {
-      // handle created events
-      if (event === "created") {
-        console.log("EmpRegistry found a new contract", JSON.stringify(data));
-        services.emps.updateAll([data.address], data.startBlock, data.endBlock).catch(console.error);
-      }
-    });
-
-    async function getLatestBlockNumber() {
-      const block = await provider.getBlock("latest");
-      return block.number;
+  // listen for new lsp contracts since after we have started api, and make sure they get state updated asap
+  // These events should only be bound after startup, since initialization above takes care of updating all contracts on startup
+  // Because there is now a event driven dependency, the lsp creator and lsp state updater must be in same process
+  serviceEvents.on("multiLspCreator", (event, data) => {
+    // handle created events
+    if (event === "created") {
+      console.log("LspCreator found a new contract", JSON.stringify(data));
+      services.lsps.updateLsps([data.address], data.startBlock, data.endBlock).catch(console.error);
     }
+  });
 
-    const newContractBlockTick = BlockInterval(detectNewContracts, initBlock.number);
-    const updateContractStateTick = BlockInterval(updateContractState, initBlock.number);
+  // listen for new emp contracts since after we start api, make sure they get state updates asap
+  serviceEvents.on("empRegistry", (event, data) => {
+    // handle created events
+    if (event === "created") {
+      console.log("EmpRegistry found a new contract", JSON.stringify(data));
+      services.emps.updateAll([data.address], data.startBlock, data.endBlock).catch(console.error);
+    }
+  });
 
-    // detect contract loop
-    utils.loop(async () => {
-      const end = profile("Detecting New Contracts");
-      // adding in a timeout rejection if the update takes too long
-      await expirePromise(
-        async () => {
-          const { startBlock, endBlock } = await newContractBlockTick(await getLatestBlockNumber());
-          console.log("Checked for new contracts between blocks", startBlock, endBlock);
-          // error out if this fails to complete in 5 minutes
-        },
-        5 * 60 * 1000,
-        "Detecting new contracts timed out"
-      )
-        .catch(console.error)
-        .finally(end);
-    }, detectContractsUpdateRateS * 1000);
-
-    // main update loop for all state, executes immediately and waits for updateRateS
-    utils.loop(async () => {
-      const end = profile("Running contract state updates");
-      // adding in a timeout rejection if the update takes too long
-      await expirePromise(
-        async () => {
-          const { startBlock, endBlock } = await updateContractStateTick(await getLatestBlockNumber());
-          console.log("Updated Contract state between blocks", startBlock, endBlock);
-          // throw an error if this fails to process in 15 minutes
-        },
-        15 * 60 * 1000,
-        "Contract state updates timed out"
-      )
-        .catch(console.error)
-        .finally(end);
-    }, updateRateS * 1000);
-
-    // coingeckos prices don't update very fast, so set it on an interval every few minutes
-    utils.loop(async () => {
-      const end = profile("Update all prices");
-      await updatePrices().catch(console.error).finally(end);
-    }, priceUpdateRateS * 1000);
+  async function getLatestBlockNumber() {
+    const block = await provider.getBlock("latest");
+    return block.number;
   }
+
+  const newContractBlockTick = BlockInterval(detectNewContracts, initBlock.number);
+  const updateContractStateTick = BlockInterval(updateContractState, initBlock.number);
+
+  // detect contract loop
+  utils.loop(async () => {
+    const end = profile("Detecting New Contracts");
+    // adding in a timeout rejection if the update takes too long
+    await expirePromise(
+      async () => {
+        const { startBlock, endBlock } = await newContractBlockTick(await getLatestBlockNumber());
+        console.log("Checked for new contracts between blocks", startBlock, endBlock);
+        // error out if this fails to complete in 5 minutes
+      },
+      5 * 60 * 1000,
+      "Detecting new contracts timed out"
+    )
+      .catch(console.error)
+      .finally(end);
+  }, detectContractsUpdateRateS * 1000);
+
+  // main update loop for all state, executes immediately and waits for updateRateS
+  utils.loop(async () => {
+    const end = profile("Running contract state updates");
+    // adding in a timeout rejection if the update takes too long
+    await expirePromise(
+      async () => {
+        const { startBlock, endBlock } = await updateContractStateTick(await getLatestBlockNumber());
+        console.log("Updated Contract state between blocks", startBlock, endBlock);
+        // throw an error if this fails to process in 15 minutes
+      },
+      15 * 60 * 1000,
+      "Contract state updates timed out"
+    )
+      .catch(console.error)
+      .finally(end);
+  }, updateRateS * 1000);
+
+  // coingeckos prices don't update very fast, so set it on an interval every few minutes
+  utils.loop(async () => {
+    const end = profile("Update all prices");
+    await updatePrices().catch(console.error).finally(end);
+  }, priceUpdateRateS * 1000);
 };