+++ packages/bot-strategy-runner/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/UMAprotocol/protocol/compare/@uma/bot-strategy-runner@1.2.2...@uma/bot-strategy-runner@1.2.3) (2021-10-08)
+
+**Note:** Version bump only for package @uma/bot-strategy-runner
+
 ## [1.2.2](https://github.com/UMAprotocol/protocol/compare/@uma/bot-strategy-runner@1.2.1...@uma/bot-strategy-runner@1.2.2) (2021-10-01)
 
 **Note:** Version bump only for package @uma/bot-strategy-runner